adjustments after rebase
