+++ src/test/scala/higherkindness/skeuomorph/avro/AvroSchemaSpec.scala
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-package skeuomorph
-package protobuf
+package higherkindness.skeuomorph.avro
 
+import higherkindness.skeuomorph.instances._
 import org.typelevel.discipline.specs2.Discipline
 import cats.laws.discipline.{FoldableTests, FunctorTests, TraverseTests}
 import cats.implicits._
 import org.specs2._
-import instances._
 
-class ProtoSchemaSpec extends Specification with ScalaCheck with Discipline {
+class AvroSchemaSpec extends Specification with ScalaCheck with Discipline {
 
   def is = s2"""
   $traverse
@@ -31,8 +30,7 @@ class ProtoSchemaSpec extends Specification with ScalaCheck with Discipline {
   $foldable
   """
 
-  val traverse =
-    checkAll("Traverse[ProtobufF]", TraverseTests[ProtobufF].traverse[Int, Int, Int, Set[Int], Option, Option])
-  val functor  = checkAll("Functor[ProtobufF]", FunctorTests[ProtobufF].functor[Int, Int, String])
-  val foldable = checkAll("Foldable[ProtobufF]", FoldableTests[ProtobufF].foldable[Int, Int])
+  val traverse = checkAll("Traverse[AvroF]", TraverseTests[AvroF].traverse[Int, Int, Int, Set[Int], Option, Option])
+  val functor  = checkAll("Functor[AvroF]", FunctorTests[AvroF].functor[Int, Int, String])
+  val foldable = checkAll("Foldable[AvroF]", FoldableTests[AvroF].foldable[Int, Int])
 }
